+++ app/src/main/java/com/sanaa/tudee_assistant/presentation/designSystem/component/SnackBar.kt
@@ -24,7 +24,7 @@ import com.sanaa.tudee_assistant.presentation.designSystem.theme.Theme
 import com.sanaa.tudee_assistant.presentation.designSystem.theme.TudeeTheme
 import com.sanaa.tudee_assistant.presentation.model.SnackBarState
 import com.sanaa.tudee_assistant.presentation.model.SnackBarStatus
-import com.sanaa.tudee_assistant.presentation.utils.dropShadow
+import com.sanaa.tudee_assistant.presentation.modifire.dropShadow
 
 @Composable
 fun SnackBar(
