+++ app/src/main/java/com/sanaa/tudee_assistant/presentation/designSystem/component/UploadBox.kt
@@ -37,7 +37,7 @@ import com.sanaa.tudee_assistant.R
 import com.sanaa.tudee_assistant.presentation.composable.CategoryThumbnail
 import com.sanaa.tudee_assistant.presentation.designSystem.theme.Theme
 import com.sanaa.tudee_assistant.presentation.designSystem.theme.TudeeTheme
-import com.sanaa.tudee_assistant.presentation.utils.drawDashedBorder
+import com.sanaa.tudee_assistant.presentation.modifire.drawDashedBorder
 
 @Composable
 fun UploadBox(
